+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.2.1](https://github.com/slnc/ifchange/compare/v0.2.0...v0.2.1) (2026-03-15)
+
+
+### Bug Fixes
+
+* exclude directive lines from co-change trigger range ([#25](https://github.com/slnc/ifchange/issues/25)) ([b6b314b](https://github.com/slnc/ifchange/commit/b6b314bbe5e243a28b743dd67b2db30a1bfdb1fb))
+
+## [0.2.0](https://github.com/slnc/ifchange/compare/v0.1.0...v0.2.0) (2026-03-10)
+
+
+### Features
+
+* add repo-root-relative paths support in ThenChange ([#21](https://github.com/slnc/ifchange/issues/21)) ([bb48538](https://github.com/slnc/ifchange/commit/bb485382ef859b87545c0a349df3a8d0fd01bce2))
+
 ## 0.1.0 (2026-03-07)